Bill S3429 version original - William Weber
BILL NUMBER: S3429 SPONSOR: WEBER TITLE OF BILL: An act to amend the penal law, in relation to enacting "Lindsey's law" PURPOSE: To strengthen the current laws on sexual harassment and retaliation in order to protect victims who comes forward. SUMMARY OF PROVISIONS: Section 1: provides the title of the legislation Section 2: amends the penal law to add section 195.