Bill S3403 version original - Neil D. Breslin
BILL NUMBER: S3403 SPONSOR: BRESLIN TITLE OF BILL: An act to amend the insurance law, in relation to modifying the interest rate within the standard nonforfeiture law for annuities PURPOSE: To ensure the flexibility and availability of innovative annuity options for consumers by addressing problems associated with the sustained low interest rate environment and the nonforfeiture provisions of the insur- ance law. SUMMARY OF PROVISIONS: Section 1.