Bill S3534 version original - Rachel May
BILL NUMBER: S3534 SPONSOR: MAY TITLE OF BILL: An act to amend the alcoholic beverage control law, in relation to a license to sell liquor at retail for consumption on certain premises PURPOSE OR GENERAL IDEA OF BILL: The bill creates an exception to the Alcoholic Beverage Control Law, allowing a license to be issued to a location currently too close to a house of worship. The license application has the support of that house of worship.