Bill S3419A version original - John Mannion
BILL NUMBER: S3419 SPONSOR: MANNION TITLE OF BILL: An act to establish the bottle redemption fraud task force; and provid- ing for the repeal of such provisions upon the expiration thereof PURPOSE: The purpose of this bill is to conduct a study to evaluate the fraud, enforcement, and reporting requirements of the Returnable Container Act. SUMMARY OF PROVISIONS: Section 1 of this bill would establish the "bottle redemption fraud task force".