Bill S3489A version original - Alexis Weik
BILL NUMBER: S3489 SPONSOR: WEIK TITLE OF BILL: An act in relation to authorizing the assessor of the town of Islip, county of Suffolk, to accept an application for exemption from real property taxes from the East Islip Fire District PURPOSE: To authorize the assessor of the town of Islip, county of Suffolk, to accept an application for exemption from real property taxes from the East Islip Fire District. SUMMARY OF PROVISIONS: Section 1.