Bill S3543 version original - Neil D. Breslin
BILL NUMBER: S3543 SPONSOR: BRESLIN TITLE OF BILL: An act to amend the insurance law, in relation to the approval of policy forms SUMMARY OF PROVISIONS: Section one amends section 3201 of the Insurance Law to add a provision which identifies certain group policies, group contracts and certif- icates issued thereunder, exempting them from the filing and approval requirements. in the section.