Bill S3496 version original - Timothy M. Kennedy
BILL NUMBER: S3496 SPONSOR: KENNEDY TITLE OF BILL: An act to amend the tax law, in relation to the amount of credit for cider, wine, and liquor under the alcoholic beverage production credit SUMMARY OF PROVISIONS: Section 1: Maintains the credit amount for: For the first 500,000 gallons of: i. beer produced in this state in the taxable year, the credit shall equal 14 cents per gallon. Updates the credit amount for the following: For the first 500,000 gallons of: ii.