Bill S3478 version original - Toby Ann Stavisky
BILL NUMBER: S3478 SPONSOR: STAVISKY TITLE OF BILL: An act to amend the administrative code of the city of New York, in relation to providing certain elderly individuals with an exemption from increases in water rates PURPOSE OR GENERAL IDEA OF BILL: The purpose of the bill is to provide water relief to elderly individ- uals in New York City. SUMMARY OF SPECIFIC PROVISIONS: Section 1. Amends the administrative code of the city of New York by adding a new section 24-340.