Bill S3430 version original - Zellnor Myrie
BILL NUMBER: S3430 SPONSOR: MYRIE TITLE OF BILL: An act to amend the election law, in relation to the representation of newly formed political parties PURPOSE: This bill would set forth the guidelines under which a new political party may form. SUMMARY OF PROVISIONS: Section 1 amends subdivision 1 of section 7-116 of the election law which addresses the placement of candidates' names on the ballot.