Bill S3446 version original - Pamela Helming
BILL NUMBER: S3446 SPONSOR: HELMING TITLE OF BILL: An act to amend the tax law, in relation to enacting "Averyana's law" PURPOSE: To establish a tax credit for buying and installing photoelectric smoke detectors. SUMMARY OF PROVISIONS: Section 1: Cites the title of act to be known as Averyana's Law. Section 2: Adds a new Subsection ccc to Section 606 of the Tax Law providing a credit to resident owners who purchase and install photoelectric detectors.