Bill S3528A version original - Nathalia Fernandez
BILL NUMBER: S3528 SPONSOR: FERNANDEZ TITLE OF BILL: An act to amend the mental hygiene law, in relation to setting standards for addiction professionals PURPOSE OR GENERAL IDEA OF BILL: Relates to setting standards for addiction professionals. SUMMARY OF PROVISIONS: Section 1 amends section 19.07 of the mental hygiene law. Subdivision (a) is amended to modify who would be provided services as a person who have or are at risk of an addictive disorder.