Bill S3391 version original - James Tedisco
BILL NUMBER: S3391 SPONSOR: TEDISCO TITLE OF BILL: An act to amend the penal law, in relation to sentence of imprisonment for the offense of predatory sexual assault against a child PURPOSE OR GENERAL IDEA OF BILL: To require a 25 year minimum prison sentence upon conviction of the crime of predatory sexual assault against a child. SUMMARY OF SPECIFIC PROVISIONS: Section 1: Amends subparagraph (ii) of paragraph (a) of subdivision 3 of section 70.