Bill S3440 version original - James Skoufis
BILL NUMBER: S3440 SPONSOR: SKOUFIS TITLE OF BILL: An act to amend the tax law, in relation to franchise tax on banking corporations PURPOSE: Reinstates a bank tax based on the highest of four bases: a tax on allo- cated entire net income, a tax on allocated alternative entire net income, a tax on allocated taxable assets, or a fixed dollar minimum tax. Banks are not allowed to segregate their income and capital into business and investment varieties.