Bill S3375 version original - James Skoufis
BILL NUMBER: S3375 SPONSOR: SKOUFIS TITLE OF BILL: An act to direct the department of state to study, evaluate, and make recommendations concerning studies directed by the legislature to be completed PURPOSE: To provide an accounting, and to determine the impact of the various "Study Bills" passed into law within the last 5 years. SUMMARY OF PROVISIONS: Section 1: Directs the Department of State to conduct a study of studies directed by the legislature to be completed over the past five years.