Bill S3523 version original - Nathalia Fernandez
BILL NUMBER: S3523 SPONSOR: FERNANDEZ TITLE OF BILL: An act to amend the education law, in relation to the practice of acupuncture and certain exemptions PURPOSE OR GENERAL IDEA OF BILL: To allow thoose who otherwise do not possess the credentials or quali- fications for the practice of acupuncture prescribed by the article or regulations promulgated hereunder or any other law but may be trained by the National Acupuncture Detoxification Association (NADA). SUMMARY OF PROVISIONS: Section 1.