Bill S3412 version original - Toby Ann Stavisky
BILL NUMBER: S3412 SPONSOR: STAVISKY TITLE OF BILL: An act to amend the education law, in relation to the registration of new curricula or programs of study offered by a not-for-profit college or university PURPOSE OR GENERAL IDEA OF BILL:: This bill would amend the education law to create an expedited program approval process for programs and institutions that meet certain crite- ria. SUMMARY OF SPECIFIC PROVISIONS:: Section 1.