Bill S3476 version original - Toby Ann Stavisky
BILL NUMBER: S3476 SPONSOR: STAVISKY TITLE OF BILL: An act to amend the administrative code of the city of New York, in relation to the filing of deeds containing restrictive covenants PURPOSE OR GENERAL IDEA OF BILL: The purpose of the bill is to allow homeowners and homeowner/neighborhood associations whose properties are subject to restrictive covenants or other types of deed restrictions, to register such deeds with the Department of Buildings. SUMMARY OF SPECIFIC PROVISIONS: Section 1.