Bill S3474A version original - Toby Ann Stavisky
BILL NUMBER: S3474REVISED 3/18/24 SPONSOR: STAVISKY TITLE OF BILL: An act to amend the education law, in relation to increasing the income eligibility threshold for the tuition assistance program PURPOSE OR GENERAL IDEA OF BILL: The purpose of this bill is to increase income eligibility threshold of the tuition assistance program for dependent students. SUMMARY OF SPECIFIC PROVISIONS: Section 1.