Bill S3550 version original - Neil D. Breslin
BILL NUMBER: S3550 SPONSOR: BRESLIN TITLE OF BILL: An act to amend the insurance law, in relation to providing for the use of treatment guidelines under the comprehensive motor vehicle repara- tions act PURPOSE: To eliminate fraudulent and unnecessary medical treatment in the no-fault system. SUMMARY OF PROVISIONS: Section 1 of this bill would require the establishment of treatment guidelines for medical treatment provided under the no-fault system.