Bill S3415 version original - Zellnor Myrie
BILL NUMBER: S3415 SPONSOR: MYRIE TITLE OF BILL: An act to amend the correction law and the executive law, in relation to prohibiting colleges from asking about an applicant's prior arrests or convictions during the pre-admissions process PURPOSE:: To make the aspirations of Brown v. Board of Education a reality for people with past criminal justice involvement and to benefit New York State's economy.