Bill S3470 version original - Gustavo Rivera
BILL NUMBER: S3470 SPONSOR: RIVERA TITLE OF BILL: An act to amend the public health law, in relation to reimbursement rates for certain services PURPOSE: The purpose of this bill is to provide reimbursement rates for psychoth- erapy services. SUMMARY OF SPECIFIC PROVISIONS: Section 1 amends clause (C) of subparagraph (ii) of paragraph (f) of subdivision 2-a of section 2807 of the public health law as it relates to psychotherapy services.