Bill S3420 version original - Thomas F. O'Mara
BILL NUMBER: S3420 SPONSOR: O'MARA TITLE OF BILL: An act to amend the penal law, in relation to stolen firearms, shotguns and rifles; and to repeal certain provisions of such law relating there- to PURPOSE: To enhance penalties for offenders who commit crimes using stolen firearms. SUMMARY OF PROVISIONS: Section 1 adds possession of a stolen firearm, rifle or shotgun to Crim- inal Possession of a Weapon in the First Degree, PL § 265.04.