Bill S3579 version original - Toby Ann Stavisky
BILL NUMBER: S3579 SPONSOR: STAVISKY TITLE OF BILL: An act to amend the education law, in relation to requiring certain text-books to reference the Sea of Japan as also the East Sea PURPOSE OR GENERAL IDEA OF BILL: To require that the Sea of Japan also be referenced as the East Sea in textbooks used in New York. SUMMARY OF SPECIFIC PROVISIONS: Section 1.