Bill S3349A version A - Nathalia Fernandez
BILL NUMBER: S3349A SPONSOR: FERNANDEZ TITLE OF BILL: An act to amend the mental hygiene law, in relation to the creation of a certified recovery residences task force PURPOSE OR GENERAL IDEA OF BILL: The purpose of the bill is to create a task force charged with estab- lishing best practice guidelines for recovery living residences. SUMMARY OF PROVISIONS: Section one adds a new section. 19.04 to the mental hygiene law, estab- lishing the recovery living task force.