Bill S3584 version original - Patrick M. Gallivan
BILL NUMBER: S3584 SPONSOR: GALLIVAN TITLE OF BILL: An act to amend the penal law, in relation to murder in the first degree PURPOSE OF BILL: This bill elevates certain enumerated sexual offenses to murder in the 1st degree when the perpetrator causes the death of the victim and such victim is found to have been given an illegal controlled substance or prescription drug to facilitate such sex offense. SUMMARY OF SPECIFIC PROVISIONS: Section 1 adds a new subdivision 1-a to Penal law section 125.