Bill S3582 version original - Michelle Hinchey
BILL NUMBER: S3582 SPONSOR: HINCHEY TITLE OF BILL: An act to amend the tax law, in relation to the treatment of excess tax credits for the rehabilitation of historic barns PURPOSE: To provide New Yorkers with an income below $60,000 a refund for excess tax credits from the rehabilitation of historic barns. SUMMARY OF PROVISIONS: Section one amends the tax law to provide New Yorkers with an income below $60,000 with a refund for excess tax credits from the rehabili- tation of historic barns.