Bill S3556 version original - Neil D. Breslin
BILL NUMBER: S3556 SPONSOR: BRESLIN TITLE OF BILL: An act to amend the penal law and the criminal procedure law, in relation to criminalizing acting as a runner or soliciting or employing a runner to procure patients or clients PURPOSE: Creates the crime of unlawful procurement of clients in three degrees. Criminalizes acting as a "runner" or soliciting or employing a "runner" to procure patients or clients. SUMMARY OF PROVISIONS: Section 1 adds three new subdivisions to section 176.