Bill S3573 version original - James Skoufis
BILL NUMBER: S3573 SPONSOR: SKOUFIS TITLE OF BILL: An act to amend the education law, in relation to the compulsory age for full time instruction and certain related exceptions PURPOSE: This bill lowers the compulsory age of full-time school attendance to five years of age, from the current age of six. It requires that all children between the ages of five to sixteen years of age attend school full-time. SUMMARY OF PROVISIONS: Section 1: Defines the term "child of compulsory school age.