Bill S3562 version original - Liz Krueger
BILL NUMBER: S3562 SPONSOR: KRUEGER TITLE OF BILL: An act to amend the tax law, in relation to the deduction for stock options PURPOSE: This legislation would specify that the deduction can only be based on the price of the stock at the time of issuance of the option, not the price of the stock when sold. SUMMARY OF PROVISIONS: The Tax Law is amended by adding a new paragraph 28 to paragraph (b) of subdivision 9 of section 208.