Bill S3565A version original - Cordell Cleare
BILL NUMBER: S3565 SPONSOR: CLEARE TITLE OF BILL: An act to amend the arts and cultural affairs law, in relation to empowering the trust for cultural resources of the city of New York to designate East Harlem as a cultural district PURPOSE OR GENERAL IDEA OF BILL: To empower the trust for cultural resources of the City of New York to designate East Harlem "El Barrio" as a cultural district. SUMMARY OF PROVISIONS: Section 1; Section 21.