Bill S3583 version original - John Mannion
BILL NUMBER: S3583 SPONSOR: MANNION TITLE OF BILL: An act to amend the environmental conservation law, in relation to advanced recycling and advanced recycling facilities PURPOSE OR GENERAL IDEA OF BILL: The purpose of this bill is to authorize the operation of advanced recy- cling facilities. SUMMARY OF PROVISIONS: Section 1 of this bill would amend section 27-0303 of the Environmental Conservation Law to clarify that certain definitions would exempt advanced recycling facilities.