Bill S3577 version original - Gustavo Rivera
BILL NUMBER: S3577 SPONSOR: RIVERA TITLE OF BILL: An act to amend the social services law, in relation to authorizing the payment of medical assistance for the provision of kidney transplants for certain residents of New York PURPOSE OR GENERAL IDEA OF BILL: This bill amends the Social Services Law. with regard to expanding available public medical assistance for kidney transplantation to include undocumented persons in New York.