Bill S3599 version original - Toby Ann Stavisky
BILL NUMBER: S3599 SPONSOR: STAVISKY TITLE OF BILL: An act to amend the penal law, in relation to creating an affirmative defense in certain cases involving sexting by persons under the age of 18 PURPOSE OR GENERAL IDEA OF BILL: This bill would provide an affirmative defense to young persons for certain acts with regard to the possessions and dissemination of such images and photographs of themselves. SUMMARY OF SPECIFIC PROVISIONS: Section 1. Amends the penal law by adding a new section 40.