Bill S3590 version original - Neil D. Breslin
BILL NUMBER: S3590 SPONSOR: BRESLIN TITLE OF BILL: An act to amend the insurance law, in relation to enacting the "automo- bile insurance fraud prevention act of 2023" PURPOSE OR GENERAL IDEA OF BILL: To reform the automobile no-fault insurance system. SUMMARY OF PROVISIONS: Section 1. This act shall be known and may be cited as the "Automobile Insurance Fraud prevention Act of 2021." Section 2. Fair claims settlement.