Bill S4049 version original - Nathalia Fernandez
BILL NUMBER: S4049 SPONSOR: FERNANDEZ TITLE OF BILL: An act to amend the state finance law, in relation to the reporting of contributions by business entities and individuals PURPOSE OR GENERAL IDEA OF BILL: The purpose of this legislation is to require the reporting of contrib- utions by business entities and individuals. SUMMARY OF PROVISIONS: Section 1. The State Finance Law is amended by adding a new section 139-1 to read as follows: § 139-1.