Bill S4081 version original - Jabari Brisport
BILL NUMBER: S4081 SPONSOR: BRISPORT TITLE OF BILL: An act to amend the social services law, in relation to assistance for guide dogs, hearing dogs and service dogs and to repeal section 303-a of such law relating thereto SUMMARY OF PROVISIONS: Section one repeals section 303-a of the social services law and adds a new section 131-y. Section two sets forth the effective date. JUSTIFICATION: Certain expenses are incurred for necessities, such as for housing and heat.