Bill S4041A version original - Shelley Mayer
BILL NUMBER: S4041 SPONSOR: MAYER TITLE OF BILL: An act to amend the education law, in relation to replacing the words handicapping conditions with the word disabilities PURPOSE: To modernize the New York Education Law to be consistent with Federal law and dignified phrasing conventions. SUMMARY OF SPECIFIC PROVISIONS: Sections 1 to 6 changes "handicapping conditions" to "disabilities". Section 2 sets the effective date.