Bill S4050 version original - Nathalia Fernandez
BILL NUMBER: S4050 SPONSOR: FERNANDEZ TITLE OF BILL: An act to amend the education law, in relation to establishing the "New York patients' right to know act" PURPOSE OR GENERAL IDEA OF BILL: To ensure that New York patients' have the right to know the necessary information and professional conduct history of the doctors that are providing them with care. SUMMARY OF PROVISIONS: Section 1 establishes the title of the act as "New York patients' right to know act.