Bill S4076 version original - Rachel May
BILL NUMBER: S4076 SPONSOR: MAY TITLE OF BILL: An act to amend the state finance law, in relation to establishing the state university of New York and the city university of New York endow- ment fund PURPOSE: The purpose of this bill is to create an endowment fund for both the state university of New York and the city university of New York with funds being held and invested by the Comptroller. SUMMARY OF PROVISIONS: Section 1.