Bill S4039 version original - Shelley Mayer
BILL NUMBER: S4039 SPONSOR: MAYER TITLE OF BILL: An act to amend the vehicle and traffic law, in relation to requiring fingerprinting as part of the criminal background check for a TNC driver PURPOSE: To ensure the safety of riders in transportation network company vehi- cles across the state by ensuring that drivers are fingerprinted. SUMMARY OF SPECIFIC PROVISIONS: Section 1.