Bill S4087 version original - Pamela Helming
BILL NUMBER: S4087 SPONSOR: HELMING TITLE OF BILL: An act to authorize Michael Telesca to take the competitive civil service examination and be placed on the eligible civil service list for employment as a full-time police officer for the county of Livingston PURPOSE: To provide the county of Livingston the opportunity to offer Mr. Teles- ca a full time position with the sheriff's department upon his satisfac- tory performance on the relevant civil service examination.