Bill S4048 version original - Nathalia Fernandez
BILL NUMBER: S4048 SPONSOR: FERNANDEZ TITLE OF BILL: An act to amend the election law, in relation to transition and inaugu- ration donations and expenses and funds for personal use PURPOSE OR GENERAL IDEA OF BILL: To create rules surrounding expenses and donations for transition and inauguration entities. SUMMARY OF PROVISIONS: Section one amends the election law to add a new article 18 entitled Transition and Inauguration Entities.