Bill S4022 version original - Joseph A. Griffo
BILL NUMBER: S4022 SPONSOR: GRIFFO TITLE OF BILL: An act to amend the vehicle and traffic law, in relation to the minimum age of eligibility for certain commercial driver's licenses PURPOSE: To lower the age of CDL licensure from 21 years of age to 18 years of age. SUMMARY OF PROVISIONS: Section 1 lowers from 21 to 18 the age at which an individual may obtain a CDL license. Section 2 marks the effective date.