Bill S4124 version original - Andrew J. Lanza
BILL NUMBER: S4124 SPONSOR: LANZA TITLE OF BILL: An act to amend the retirement and social security law, in relation to increasing the retiree public earnings cap PURPOSE: This bill would increase the amount that a public retiree under'sixtyfive years of age may earn in public employment without a reduction of his or her retirement benefits. SUMMARY OF PROVISIONS: The bill amends subdivision 2 of Section 212 of the Retirement and Social Security Law.