Bill S4092 version original - Andrew J. Lanza
BILL NUMBER: S4092 SPONSOR: LANZA TITLE OF BILL: An act to amend the education law, in relation to teen dating violence education programs PURPOSE: To promote education of students regarding teen dating violence. SUMMARY OF PROVISIONS: Section one amends subdivision 7 of section 11 of the education law, as amended by chapter 102 of the laws of 2012, to include teen dating violence within the definition of harassment and bullying.