Bill S4130 version original - John Mannion
BILL NUMBER: S4130 SPONSOR: MANNION TITLE OF BILL: An act to amend the general business law, in relation to availability for sale of advertised merchandise PURPOSE: The purpose of this bill is to require retailers to provide a raincheck to purchasers for the sale of advertised merchandise. SUMMARY OF PROVISIONS: Section 1 of this bill would amend the General Business Law to require retailers to provide a raincheck to purchasers for the sale of adver- tised merchandise under certain conditions.