Bill S4166 version original - Brad Hoylman-Sigal
BILL NUMBER: S4166 SPONSOR: HOYLMAN-SIGAL TITLE OF BILL: An act to amend the executive law, in relation to increasing the fees certain notaries public shall be entitled to PURPOSE: To increase the notary fee from $2.00 to $5.00 in the State of New York. SUMMARY OF PROVISIONS: Section 136 of the Executive Law is amended to increase the notarial fees a notary public shall be entitled to from $2.00 to $5.00.