Bill S4096 version original - Andrew J. Lanza
BILL NUMBER: S4096 SPONSOR: LANZA TITLE OF BILL: An act to amend the education law, in relation to random testing for anabolic steroids in athletes in public and private schools; and making an appropriation therefor PURPOSE: This bill provides for random testing for anabolic steroids in athletes in public and private schools. SUMMARY OF PROVISIONS: Section 1. The education law is amended by adding a new article 20 to Section 949 stating the Legislative intent.