Bill S4150 version original - James Sanders Jr.
BILL NUMBER: S4150 SPONSOR: SANDERS TITLE OF BILL: An act to amend the general business law, in relation to limiting the acquisition of a rifle or shotgun to one per ninety-day period PURPOSE: To strengthen laws in the state to prevent diversion of guns to the illegal marketplace. SUMMARY OF PROVISIONS: Section 1 states the legislative findings.