Bill S4143 version original - James Sanders Jr.
BILL NUMBER: S4143 SPONSOR: SANDERS TITLE OF BILL: An act to amend the family court act, in relation to service of summons- es upon parents in juvenile delinquency cases PURPOSE: To better engage the families of juveniles involved in juvenile delin- quency proceedings and to assist the Family Court in formulating dispo- sitions in the best interest of juveniles and communities. SUMMARY OF PROVISIONS: Section 312.