Bill S4163 version original - Brad Hoylman-Sigal
BILL NUMBER: S4163 SPONSOR: HOYLMAN-SIGAL TITLE OF BILL: An act to amend the insurance law, in relation to prohibiting insurers from cancelling, refusing to issue or renew renters' insurance based on the breed of dog owned PURPOSE: This bill prohibits insurers from cancelling or refusing to issue a renters or homeowners insurance policy based on the ownership of a specific breed of dog. SUMMARY OF SPECIFIC PROVISIONS: Section one amends the insurance law is amended by adding a new section 3421.