Bill S4138 version original - James Sanders Jr.
BILL NUMBER: S4138REVISED 2/10/23 SPONSOR: SANDERS TITLE OF BILL: An act to amend the election law, in relation to providing the option for candidates for office and petition witnesses not to put their address on designation petitions PURPOSE OR GENERAL IDEA OF BILL: The bill provides the option for candidates for local and state office and petition witnesses not to put their address on designation petitions. This bill is called the "Esther Salas Candidate Privacy Act.