Bill S4110 version original - Andrew J. Lanza
BILL NUMBER: S4110 SPONSOR: LANZA TITLE OF BILL: An act to amend the penal law, in relation to making rape in the first degree a class A-I felony with a penalty of life imprisonment without parole PURPOSE: Provides that rape in the first degree shall be a class A-I felony and shall be punishable by life in prison without parole. SUMMARY OF PROVISIONS: Section 1 amends the closing paragraph of section 130.35 of the penal law, as amended by chapter 1 of the laws of 2000.