Bill S4149 version original - Anthony H. Palumbo
BILL NUMBER: S4149 SPONSOR: PALUMBO TITLE OF BILL: An act to amend the domestic relations law and the family court act, in relation to prohibiting forensic evaluations in a custody or visitation proceeding; and to repeal certain provisions of the domestic relations law relating thereto PURPOSE: To protect children by prohibiting forensic evaluations in child custody and visitation proceedings. SUMMARY OF PROVISIONS: Section 1.