Bill S4122 version original - Andrew J. Lanza
BILL NUMBER: S4122 SPONSOR: LANZA TITLE OF BILL: An act to amend the penal law, in relation to making conviction of assault of a social worker a felony PURPOSE: Provides that conviction of assault on a social worker which prevents him or her from performing a lawful duty is a felony SUMMARY OF PROVISIONS: Section one. Section 120.