Bill S4089 version original - Andrew J. Lanza
BILL NUMBER: S4089 SPONSOR: LANZA TITLE OF BILL: An act to amend the penal law, in relation to non-support of a child PURPOSE: Requires a delinquent parent to demonstrate as an affirmative defense that the inability to pay excuses his or her failure to provide child support for their minor child. SUMMARY OF PROVISION: Section one amends sections 260.05 & 260.