Bill S4105 version original - Andrew J. Lanza
BILL NUMBER: S4105 SPONSOR: LANZA TITLE OF BILL: An act to amend the agriculture and markets law, in relation to penal- ties for causing the death of an animal PURPOSE: The purpose of this bill is to elevate the penalty for killing an animal while exhibiting cruelty toward that animal. SUMMARY OF PROVISIONS: Section 1 adds § 352 to the agriculture and markets law which creates the offense of causing the death of an animal.