Bill S4132 version original - John Mannion
BILL NUMBER: S4132 SPONSOR: MANNION TITLE OF BILL: An act to amend the civil service law, in relation to preventing outsourcing during a hiring freeze PURPOSE: The purpose of this bill is to prevent outsourcing during a hiring freeze. SUMMARY OF PROVISIONS:- Section 1 of this bill would establish that no entity under the control of the Executive shall contract, subcontract, or hire any third party during a hiring freeze.