Bill S4131 version original - John Mannion
BILL NUMBER: S4131 SPONSOR: MANNION TITLE OF BILL: An act to amend the tax law, in relation to personal income tax credit for residential solar energy systems PURPOSE: The purpose of this bill is to expand eligibility for personal income tax credit for residential solar energy systems. SUMMARY OF PROVISIONS: Section 1 of this bill would amend subparagraph (A) of paragraph 2 of subsection (g-1) of section 606 of the Tax Law to allow solar tax cred- its on one residence per homeowner.