Bill S4154 version original - James Sanders Jr.
BILL NUMBER: S4154 SPONSOR: SANDERS TITLE OF BILL: An act to amend the penal law, in relation to photographing or recording a law enforcement officer PURPOSE: To establish that a person is permitted to exercise their right to photograph or record a law enforcement officer. SUMMARY OF PROVISIONS: Section 1 amends section 195.