Bill S4162 version original - Jessica Ramos
BILL NUMBER: S4162 SPONSOR: RAMOS TITLE OF BILL: An act to amend the civil rights law, in relation to imposing limita- tions on the use of drones for law enforcement purposes PURPOSE OR GENERAL IDEA OF BILL: This legislation will impose limitations on the use of drones for law enforcement purposes. SUMMARY OF PROVISIONS: Section 1: This act shall be known and may be cited as the "protect our privacy (POP) act".