Bill S4151 version original - James Sanders Jr.
BILL NUMBER: S4151 SPONSOR: SANDERS TITLE OF BILL: An act to amend the public health law, in relation to establishing the infant vision information, education and wellness program PURPOSE: To enact the Infant Vision Information, Education and Wellness Program for screening newborns and infants for vision abnormalities. SUMMARY OF PROVISIONS: Section 1 provides this act shall be known and may be cited as the "Infant Vision Information, Education and Wellness Act.