Bill S4090 version original - Andrew J. Lanza
BILL NUMBER: S4090 SPONSOR: LANZA TITLE OF BILL: An act to amend the penal law, in relation to establishing the offense of official interference with transportation PURPOSE: Establishes the class E felony of official interference with transporta- tion. SUMMARY OF PROVISIONS: Section one of the bill amends the penal law by adding a new section 195.03. Section two is the effective date.