Bill S4219 version original - Joseph A. Griffo
BILL NUMBER: S4219 SPONSOR: GRIFFO TITLE OF BILL: An act to amend the tax law, in relation to providing tax credits for installing wheelchair ramps and stair lifts into a residence PURPOSE: This bill would provide a tax credit so that adding accessibility equip- ment in a home with persons with mobility issues will be more afforda- ble. SUMMARY OF PROVISIONS: Section 1: Allows for a tax credit to be applied in the amount of fifty percent when durable medical equipment is purchased.