Bill S4227 version original - Joseph A. Griffo
BILL NUMBER: S4227 SPONSOR: GRIFFO TITLE OF BILL: An act to amend the public health law, in relation to the maximum amount of individual awards for the physician loan repayment program and physi- cian practice support program PURPOSE: To increase the amount of money that is awarded for the physician loan repayment program and the physician practice support program. SUMMARY OF PROVISIONS: Section 1 increases the amount of the Physician Loan Repayment Program award from $150,000 to $250,000.