Bill S4192 version original - Andrew J. Lanza
BILL NUMBER: S4192 SPONSOR: LANZA TITLE OF BILL: An act to amend the penal law, in relation to the seizure and forfeiture of vehicles, vessels and aircraft used in counterfeit goods PURPOSE: To provide for the seizure and forfeiture of the property used in traf- ficking in counterfeit goods. SUMMARY OF PROVISIONS: This legislation would provide for the seizure and forfeiture of the property of trademark counterfeiters and those who traffic in counter- feit goods.