Bill S4211 version original - Andrew J. Lanza
BILL NUMBER: S4211 SPONSOR: LANZA TITLE OF BILL: An act to amend the penal law and the criminal procedure law, in relation to crimes committed during a riot PURPOSE: Enhance the penalties of for the crimes of rioting and burglary or arson during a riot. Also makes the crimes of rioting bail eligible. SUMMARY OF PROVISIONS: This bill enhances the penalties for the crimes of inciting a riot and riot in the first and second degrees.