Bill S4225 version original - Joseph A. Griffo
BILL NUMBER: S4225 SPONSOR: GRIFFO TITLE OF BILL: An act to amend the penal law, in relation to the definition of wiretap- ping PURPOSE: That an individual recording a conversation or incident, where there is a reasonable expectation of privacy, must state their intentions to use a recording device so that all parties present know that such device is going to be used/is being used. SUMMARY OF PROVISIONS: Section 1 amends the definition of "wiretapping" under section 250.00 of the penal law.