Bill S4243 version original - Alexis Weik
BILL NUMBER: S4243 SPONSOR: WEIK TITLE OF BILL: An act to amend the tax law, in relation to exempting the sale of motor vehicles which will be modified for operation by, or for the transporta- tion of, a person with a disability from state sales and compensating use taxes PURPOSE: To exempt the purchase of a new or certain used vehicles that require modifications to permit use by a person with a disability. SUMMARY OF PROVISIONS: Section one includes eligibility criteria for the exemption: 1.