Bill S4226 version original - Joseph A. Griffo
BILL NUMBER: S4226 SPONSOR: GRIFFO TITLE OF BILL: An act to amend the real property tax law, in relation to providing a filing extension for the "enhanced" STAR exemption to eligible appli- cants PURPOSE: To allow a retroactive, one-time extension to filers for "enhanced" STAR who have missed the filing deadline. SUMMARY OF PROVISIONS: Section 1 allows any filer for the "enhanced" STAR program, through written request to the Commissioner, to receive an extension on their application.