Bill S4186 version original - Andrew J. Lanza
BILL NUMBER: S4186 SPONSOR: LANZA TITLE OF BILL: An act to amend the penal law, in relation to establishing the crime of feticide PURPOSE: Relates to establishing the crime of feticide SUMMARY OF PROVISIONS: Section one amends the penal law by adding section 125.24 to create the crime of Feticide. Feticide is the termination of one's pregnancy with the intent to cause the death of an unborn child and shall be a Class A felony.