Bill S4185 version original - Andrew J. Lanza
BILL NUMBER: S4185 SPONSOR: LANZA TITLE OF BILL: An act to amend the vehicle and traffic law, in relation to establishing a yellow dot program for persons fifty-five years of age or older PURPOSE: Directs the commissioner of motor vehicles to establish a yellow dot program under which vehicle occupants' medical information is available in the glove compartment. SUMMARY OF SPECIFIC PROVISIONS: Section one amends section 200 of the vehicle and traffic law by adding a new subdivision three.