Bill S4217 version original - Andrew J. Lanza
BILL NUMBER: S4217 SPONSOR: LANZA TITLE OF BILL: An act to enact "Michelle and Jordan's law"; and to amend the vehicle and traffic law, in relation to speed contests and races PURPOSE: Establishes Michelle and Jordan's law and relates to the penalties for participating in an unlawful speed contest or race. SUMMARY OF PROVISIONS: Section 1. Short title. This act shall be known and may be cited as "Michelle and Jordan's Law." Section 2.