Bill S4181 version original - Andrew J. Lanza
BILL NUMBER: S4181 SPONSOR: LANZA TITLE OF BILL: An act to amend the public health law and the education law, in relation to generic drug products; and to repeal paragraph (o) of subdivision 1 of section 206 of the public health law relating thereto PURPOSE OR GENERAL IDEA OF BILL: To make relevant information related to generic drugs freely and public- ly available. SUMMARY OF SPECIFIC PROVISIONS: Section one repeals paragraph (o) of subdivision 1 of section 206 of the public health law (PHL).