Bill S4207 version original - Andrew J. Lanza
BILL NUMBER: S4207 SPONSOR: LANZA TITLE OF BILL: CONCURRENT RESOLUTION OF THE SENATE AND ASSEMBLY proposing an amendment to the constitution, in relation to eligibility to run for president of the United States PURPOSE OR GENERAL IDEA OF BILL: Relates to eligibility requirements to run for the office of president of the United States. SUMMARY OF SPECIFIC PROVISIONS: Section one amends the state constitution by adding a new article XIII-A, Running for Public Office.