Bill S4193 version original - James Sanders Jr.
BILL NUMBER: S4193 SPONSOR: SANDERS TITLE OF BILL: An act to direct the New York state department of health to conduct a study on the incidences of cancer clusters in cities and towns having a population of more than ninety thousand PURPOSE: To require the New York State Department of Health to conduct a study on the incidences of cancer clusters in cities and towns having a popu- lation over 90,000 people. The study should be completed by September 30, 2023. SUMMARY OF PROVISIONS: § 1.