Bill S4188 version original - Andrew J. Lanza
BILL NUMBER: S4188 SPONSOR: LANZA TITLE OF BILL: An act to amend the criminal procedure law and the executive law, in relation to requiring that all family members of a crime victim and all interested parties who want to give a victim impact statement to parole board members be allowed to do so PURPOSE: Provides for family members of a crime victim and interested parties to make statements to members of the parole board. SUMMARY OF PROVISIONS: Section one amends Section 440.