Bill S4221 version original - Joseph A. Griffo
BILL NUMBER: S4221 SPONSOR: GRIFFO TITLE OF BILL: An act to amend the penal law, in relation to the definition of wiretap- ping PURPOSE: The purpose of this bill is to require that anytime a recording device is used during a conversation or incident, where there is a reasonable expectation of privacy, that all parties consent to the recording of such conversation or incident. SUMMARY OF PROVISIONS: Section 1 changes the definition of "wiretapping" under section 250.