Bill S4199 version original - James Sanders Jr.
BILL NUMBER: S4199 SPONSOR: SANDERS TITLE OF BILL: An act to amend the insurance law, in relation to homeowners insurance deductibles triggers PURPOSE: The purpose of this bill is to help promote better understanding of the applicability and amount of hurricane windstorm deductibles in homeown- ers and dwelling fire policies and to establish reasonable standards for the operation of hurricane windstorm deductibles. SUMMARY OF PROVISIONS: Section 1.