Bill S4358 version original - Nathalia Fernandez
BILL NUMBER: S4358 SPONSOR: FERNANDEZ TITLE OF BILL: An act to amend the penal law, in relation to consent PURPOSE OR GENERAL IDEA OF BILL: To add an affirmative definition of consent to the penal code. SUMMARY OF PROVISIONS: Section 1 amends the Penal Law to define consent to sexual contact as a knowing, voluntary, and mutual decision among all participants.