Bill S4290 version original - Andrew J. Lanza
BILL NUMBER: S4290 SPONSOR: LANZA TITLE OF BILL: An act to amend the education law, in relation to requiring certain public and nonpublic schools to provide transportation or reimbursement for transportation for certain students residing in an area containing at least three hundred students within a one-mile radius PURPOSE OR GENERAL IDEA OF BILL: To include certain public and nonpublic schools in the after 4:00pm program. SUMMARY OF PROVISIONS: Section 1.