Bill S4312 version original - Luis R. Sepúlveda
BILL NUMBER: S4312 SPONSOR: SEPULVEDA TITLE OF BILL: An act to amend the election law, in relation to the registration status of voters PURPOSE OR GENERAL IDEA OF BILL: This purpose of this bill is to prevent the disenfranchisement of voters through hurdles such as registration cancellation. SUMMARY OF SPECIFIC PROVISIONS: Section 1. Amends Section 17-106 of the Election law. Makes misconduct of an election official a felony.